PythonComputing/Projects/Dyslexhelp/requirements.txt
flask
opencv-python
pytesseract
numpy
google-cloud-vision
Pillow  